Tl/hero

See merge request mreinsch1/pm_dungeon!20

+++ src/items/shields/Shield.java
@@ -4,11 +4,11 @@ import items.EquipableItem;
 import main.Game;
 
 public abstract class Shield extends EquipableItem {
-  protected int defenseValue;
+  protected float defenseValue;
   public Shield(Game game) {
     super(game);
   }
-  public int getDefenseValue(){
+  public float getDefenseValue(){
     return defenseValue;
   }
 }
